Move exceptions to new folder
