+++ src/main/java/reposense/authorship/FileInfoExtractor.java
@@ -19,7 +19,7 @@ import java.util.stream.Stream;
 
 import reposense.authorship.model.FileInfo;
 import reposense.authorship.model.LineInfo;
-import reposense.git.CommitNotFoundException;
+import reposense.git.exception.CommitNotFoundException;
 import reposense.git.GitCheckout;
 import reposense.git.GitDiff;
 import reposense.git.GitRevList;
