+++ src/main/java/reposense/git/exception/GitCloneException.java
@@ -1,4 +1,4 @@
-package reposense.git;
+package reposense.git.exception;
 
 
 public class GitCloneException extends Exception {
