+++ src/main/java/reposense/model/RepoConfiguration.java
@@ -9,7 +9,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.logging.Logger;
 
-import reposense.git.BranchNotFoundException;
+import reposense.git.exception.BranchNotFoundException;
 import reposense.git.GitBranch;
 import reposense.system.LogsManager;
 import reposense.util.FileUtil;
