+++ src/test/java/reposense/authorship/FileAnalyzerTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 
 import reposense.authorship.model.FileInfo;
 import reposense.authorship.model.FileResult;
-import reposense.git.CommitNotFoundException;
+import reposense.git.exception.CommitNotFoundException;
 import reposense.git.GitCheckout;
 import reposense.model.Author;
 import reposense.model.CommitHash;
